socwall: Place dowlanded images in a separate folder
Some system (e.g. KDE) do not refresh the wallpaper
display when the file changes.
Putting the images in a dedicated folder allows to use
the diaporama feature to accomplish that.
One just have to set the refresh time of the diaporama to
match the refresh rate of the socwall calls, making the
wallpaper refresh since there is only one image in the folder.